pyproject.toml
[tool.poetry]
name = "car-classifier"
version = "0.1.0"
description = ""
authors = ["AJ <akarsh.1995.02@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.6.2,<3.7"
Keras = "^2.4.3"
tensorflow = "^2.4.0"
matplotlib = "^3.3.3"
pandas = "<2"
dataclasses = "^0.8"
Pillow = "^8.1.0"
opencv-python = "^4.5.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
